converting to long

+++ diff/src/main/scala/com/alphasystem/diff/Line.scala
@@ -1,3 +1,3 @@
 package com.alphasystem.diff
 
-case class Line[T](lineNumber: Int, text: T)
+case class Line[T](lineNumber: Long, text: T)
